Fixes a URL Parsing error reported in Issue Tracker of the product Poi #101 (->r41825, r41869)
Implements enhancements based on Alexander Limi's email 2007-05-10:
1.) Add rel="nofollow" to links
2.) Disguise email addresses with unicode to avoid email spam crawlers to pickup 
(->r41930)

Implements enhancement based on Alexander Limi's email 2007-05-16:
In case an url is too long (60 characters is the default max size), the url is
shortened by replacing a slice in the middle by an ellipsis (default:
"[&hellip;]").
We try to create a reasonable replacement by replacing full paths - like:
http://www.example.com/[...]/index.html
In case this is not possible we concatinate the beginning and the end to get a
string with the desired size. 
(->r42220, r42221)
 
Upstream checkin from https://svn.plone.org/svn/collective/intelligenttext/trunk/ r41825, 41869, 41930, 42220, 42221.


svn path=/plone.intelligenttext/trunk/; revision=15244