Recognize urls at the end of sentences
URLs do not end with punctuation marks in general.
If they do, they're most probably part of a sentence.
In that case the punctuation mark needs to be split from the URL.
This allows to embed URLs in sentences without needing to add an extra
empty space after the URL before setting the punctuation mark.

To force the punctuation mark to be part of the URL, brackets can be
used (as shown in one of the tests).